ORM system shift to the models.go file
